Harden timer error handling, defer mat allocation until arg validation, and remove dead timer stub

+++ Tuning/sample_code/12_swp/src/c/memory.c
@@ -18,7 +18,7 @@
 int *create_int_1d_array( const int n )
 {  
    if ( n <= 0 ) return NULL;
-   int nbytes = sizeof(int)*n;
+   size_t nbytes = sizeof(int)*n;
    int *array = (int *)malloc(nbytes); 
    return array;
 }
@@ -38,7 +38,7 @@ void release_int_1d_array( int *array)
 double *create_dp_1d_array( const int n )
 {  
    if ( n <= 0 ) return NULL;
-   int nbytes = sizeof(double)*n;
+   size_t nbytes = sizeof(double)*n;
    double *array = (double *)malloc(nbytes); 
    return array;
 }
@@ -58,7 +58,7 @@ void release_dp_1d_array( double *array)
 int **create_int_2d_array( const int n1, const int n2)
 {
    if ( n1 <= 0 || n2 <= 0 ) return NULL;
-   int nbytes = sizeof(int)*n1*n2; 
+   size_t nbytes = sizeof(int)*n1*n2; 
    int *data = (int *)malloc(nbytes);
    nbytes = sizeof(int *)*n1;
    int **array = (int **)malloc(nbytes);
@@ -89,7 +89,7 @@ void release_int_2d_array( int **array)
 double **create_dp_2d_array( const int n1, const int n2)
 {
    if ( n1 <= 0 || n2 <= 0 ) return NULL;
-   int nbytes = sizeof(double)*n1*n2; 
+   size_t nbytes = sizeof(double)*n1*n2; 
    double *data = (double *)malloc(nbytes);
    nbytes = sizeof(double *)*n1;
    double **array = (double **)malloc(nbytes);
@@ -120,7 +120,7 @@ void release_dp_2d_array( double **array)
 int ***create_int_3d_array( const int n1, const int n2, const int n3)
 {
    if ( n1 <= 0 || n2 <= 0 || n3 <= 0 ) return NULL;
-   int nbytes = sizeof(int)*n1*n2*n3; /* 64-bit unsigned int should be used */ 
+   size_t nbytes = sizeof(int)*n1*n2*n3; /* size_t is used for the byte count */
    int *data = (int *)malloc(nbytes);
    nbytes = sizeof(int *)*n1*n2;
    int **array2 = (int **)malloc(nbytes);
@@ -161,7 +161,7 @@ void release_int_3d_array( int ***array)
 double ***create_dp_3d_array( const int n1, const int n2, const int n3)
 {
    if ( n1 <= 0 || n2 <= 0 || n3 <= 0 ) return NULL;
-   int nbytes = sizeof(double)*n1*n2*n3; /* 64-bit unsigned int should be used */ 
+   size_t nbytes = sizeof(double)*n1*n2*n3; /* 64-bit unsigned int should be used */ 
    double *data = (double *)malloc(nbytes);
    nbytes = sizeof(double *)*n1*n2;
    double **array2 = (double **)malloc(nbytes);
